+++ CHANGELOG.md
@@ -5,11 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## unreleased
+## 0.2.0 - 2021-10-11
 
 ### Deleted
 
 * spreadsheet description from aggregated results
+* card view hook
 
 ### Added
 
@@ -22,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 * wiki formatting to spreadsheet description
+* inheritation button for accepting the current calculated result as final result
 
 ### Fixed